+++ backend/domain_types/src/router_data.rs
@@ -167,7 +167,8 @@ pub enum ConnectorSpecificAuth {
         api_key: Secret<String>,
     },
     Revolut {
-        api_key: Secret<String>,
+        secret_api_key: Secret<String>,
+        signing_secret: Option<Secret<String>>,
     },
     Shift4 {
         api_key: Secret<String>,
@@ -287,6 +288,10 @@ pub enum ConnectorSpecificAuth {
     Braintree {
         public_key: Secret<String>,
         private_key: Secret<String>,
+    },
+    Truelayer {
+        client_id: Secret<String>,
+        client_secret: Secret<String>,
     },
     Worldpay {
         username: Secret<String>,
@@ -758,7 +763,8 @@ impl ForeignTryFrom<grpc_api_types::payments::ConnectorAuth> for ConnectorSpecif
                 merchant_code: worldpayxml.merchant_code.ok_or_else(err)?,
             }),
             AuthType::Revolut(revolut) => Ok(Self::Revolut {
-                api_key: revolut.api_key.ok_or_else(err)?,
+                secret_api_key: revolut.secret_api_key.ok_or_else(err)?,
+                signing_secret: revolut.signing_secret,
             }),
             AuthType::Loonio(loonio) => Ok(Self::Loonio {
                 merchant_id: loonio.merchant_id.ok_or_else(err)?,
@@ -868,7 +874,8 @@ impl ForeignTryFrom<(&ConnectorAuthType, &connector_types::ConnectorEnum)>
             },
             ConnectorEnum::Revolut => match auth {
                 ConnectorAuthType::HeaderKey { api_key } => Ok(Self::Revolut {
-                    api_key: api_key.clone(),
+                    secret_api_key: api_key.clone(),
+                    signing_secret: None,
                 }),
                 _ => Err(err().into()),
             },
@@ -1046,6 +1053,13 @@ impl ForeignTryFrom<(&ConnectorAuthType, &connector_types::ConnectorEnum)>
                 ConnectorAuthType::BodyKey { api_key, key1 } => Ok(Self::Rapyd {
                     access_key: api_key.clone(),
                     secret_key: key1.clone(),
+                }),
+                _ => Err(err().into()),
+            },
+            ConnectorEnum::Truelayer => match auth {
+                ConnectorAuthType::BodyKey { api_key, key1 } => Ok(Self::Truelayer {
+                    client_id: api_key.clone(),
+                    client_secret: key1.clone(),
                 }),
                 _ => Err(err().into()),
             },
@@ -1656,7 +1670,6 @@ impl ErrorResponse {
 }
 
 #[derive(Debug, Clone, serde::Deserialize)]
-#[serde(rename_all = "camelCase")]
 pub struct ApplePayCryptogramData {
     pub online_payment_cryptogram: Secret<String>,
     pub eci_indicator: Option<String>,